--- a/src/methods/_getMilestone.js
+++ b/src/methods/_getMilestone.js
@@ -1,11 +1,14 @@
 Roadmap.prototype._getMilestone = function(columnIdx, milestoneRank) {
     let toReturn = null;
+    const length = this._data.milestones.length;
 
-    this._data.milestones.forEach((milestone) => {
+    for (let x = 0; x < length; x++) {
+        const milestone = this._data.milestones[x];
         if (milestone.rank === milestoneRank && milestone.belongsToColumnIdx === columnIdx) {
             toReturn = milestone;
+            break;
         }
-    });
+    }
 
     return toReturn;
 }