--- a/src/methods/_getMilestones.js
+++ b/src/methods/_getMilestones.js
@@ -1,11 +1,14 @@
 Roadmap.prototype._getMilestones = function(columnIdx) {
     let toReturn = [];
+    const length = this._data.milestones.length;
 
-    this._data.milestones.forEach((milestone) => {
+    for (let x = 0; x < length; x++) {
+        const milestone = this._data.milestones[x];
+
         if (milestone.belongsToColumnIdx === columnIdx) {
             toReturn.push(milestone);
         }
-    });
+    }
 
     return toReturn;
 }