base time: 1.61s

phase 1 - 1.58s
    - Replace _getMilestones forEach with for loop

phase 2 - 1.56s
    - Cache the milestones array length in _getMilestones loop

phase 3 - 1.55s
    - Replace _getMilestone forEach with for loop w/ cached array length

phase 4 - 1.55s
    - In _getMilestone break out of for loop early when found milestone

phase 5 - 1.55s
    - Replace milestones loop in milestones with for loop w/ cached array length
    - Replace columnRanks loop in milestones with for loop w/ cached array length

phase 6 - 1.55s
    - Replace milestones loop in markComplete with for loop w/ cached array length

phase 7 - 1.53s
    - Replace columnMilestones loop in _build with for loop w/ cached array length

phase 8 - 246.15ms
    - Improve code so that _build is only called once - Now wait for all functions to call until actually building

total saved time: 1.36385s